refactor(config): split config.ts into separate modules
- Split config.ts into metadata.ts and profile.ts for better separation of concerns
- Update import paths across components to use new config structure
- Fix metadata import path in lib/metadata.ts
