+++ app/components/HeroSection.tsx
@@ -1,6 +1,6 @@
 import { Box, Heading, Text, Flex, Avatar } from "@chakra-ui/react";
 import { ThemeConfig } from "../types";
-import { personalInfo } from "../lib/config";
+import { personalInfo } from "../config/profile";
 
 interface HeroSectionProps {
   themeConfig: ThemeConfig;
