+++ app/components/shared/Footer.tsx
@@ -1,6 +1,6 @@
 import { Box, Text } from "@chakra-ui/react";
 import { ThemeConfig } from "../../types";
-import { personalInfo } from "../../lib/config";
+import { personalInfo } from "../../config/profile";
 
 interface FooterProps {
   themeConfig: ThemeConfig;
