+++ app/config/metadata.ts
@@ -1,0 +1,18 @@
+import { SiteMetadata } from "../types";
+
+export const siteMetadata: SiteMetadata = {
+  title: "Kazuki Nagasawa - Software Engineer",
+  description:
+    "Software engineer based in Osaka, working on web application design, development, and maintenance.",
+  keywords: [
+    "Software Engineer",
+    "Web Developer",
+    "Kazuki",
+    "Nagasawa",
+    "Osaka",
+  ],
+  author: "Kazuki Nagasawa",
+  ogImage: "images/og-image.png",
+  twitterHandle: "@k_nagasawa_",
+  lang: "ja",
+};
